requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
altgraph=0.17.3=py310h06a4308_0
amqp=2.6.1=py_0
asgiref=3.5.2=py310h06a4308_0
asttokens=2.0.5=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
billiard=3.6.4.0=py310h7f8727e_0
blas=1.0=mkl
boltons=23.0.0=py310h06a4308_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
cachelib=0.9.0=py310h06a4308_0
celery=4.4.0=py_0
certifi=2023.7.22=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py310h06a4308_0
comm=0.1.2=py310h06a4308_0
conda=23.7.2=py310h06a4308_0
conda-package-handling=2.2.0=py310h06a4308_0
conda-package-streaming=0.9.0=py310h06a4308_0
contourpy=1.0.5=py310hdb19cb5_0
cryptography=39.0.1=py310h9ce1e76_0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
django=4.1=py310h06a4308_0
django-compat=1.0.15=pypi_0
django-cors-headers=4.2.0=pypi_0
django-filter=23.2=pypi_0
django4-background-tasks=1.2.7=pypi_0
djangorestframework=3.14.0=pypi_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.9=h6a678d5_0
flask=2.2.2=py310h06a4308_0
flask-session=0.4.0=py310h06a4308_0
fontconfig=2.14.1=h52c9d5c_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=he621ea3_2
greenlet=2.0.1=py310h6a678d5_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=58.2=he6710b0_3
idna=3.4=py310h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.19.2=py310h2f386ee_0
ipython=8.12.0=py310h06a4308_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jedi=0.18.1=py310h06a4308_1
jinja2=3.1.2=py310h06a4308_0
jpeg=9e=h5eee18b_1
jsonpatch=1.32=pyhd3eb1b0_0
jsonpointer=2.1=pyhd3eb1b0_0
jupyter_client=8.1.0=py310h06a4308_0
jupyter_core=5.3.0=py310h06a4308_0
kiwisolver=1.4.4=py310h6a678d5_0
kombu=4.6.3=py_0
krb5=1.20.1=h568e23c_1
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libllvm10=10.0.1=hbcb73fb_5
libllvm11=11.1.0=h9e868ea_6
libpng=1.6.39=h5eee18b_0
libpq=12.15=h37d81fd_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.0=h6a678d5_2
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
llvmlite=0.39.1=py310he621ea3_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=py310h06a4308_0
markupsafe=2.1.1=py310h7f8727e_0
matplotlib=3.7.1=py310h06a4308_1
matplotlib-base=3.7.1=py310h1128e8f_1
matplotlib-inline=0.1.6=py310h06a4308_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
munkres=1.1.4=py_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py310h06a4308_0
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numba=0.56.4=py310h1128e8f_0
numpy=1.23.5=py310hd5efca6_0
numpy-base=1.23.5=py310h8e6c178_0
openssl=1.1.1w=h7f8727e_0
packaging=23.0=py310h06a4308_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py310h6a678d5_0
pip=23.0.1=py310h06a4308_0
platformdirs=2.5.2=py310h06a4308_0
pluggy=1.0.0=py310h06a4308_1
ply=3.11=py310h06a4308_0
prompt-toolkit=3.0.36=py310h06a4308_0
psutil=5.9.0=py310h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycosat=0.6.4=py310h5eee18b_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyinstaller=5.6.2=py310h5eee18b_0
pyinstaller-hooks-contrib=2022.14=py310h06a4308_0
pyopenssl=23.0.0=py310h06a4308_0
pyotp=2.8.0=pypi_0
pyparsing=3.0.9=py310h06a4308_0
pyqt=5.15.7=py310h6a678d5_1
pyqt5-sip=12.11.0=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.10=h7a1cb2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-nmap=0.7.1=pypi_0
pytz=2022.7=py310h06a4308_0
pyzmq=23.2.0=py310h6a678d5_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.28.1=py310h06a4308_1
ruamel.yaml=0.17.21=py310h5eee18b_0
ruamel.yaml.clib=0.2.6=py310h5eee18b_1
scapy=2.4.3=py_0
setuptools=66.0.0=py310h06a4308_0
simplejson=3.17.6=py310h7f8727e_0
sip=6.6.2=py310h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlalchemy=1.4.39=py310h5eee18b_0
sqlite=3.41.2=h5eee18b_0
sqlparse=0.4.4=py310h06a4308_0
stack_data=0.2.0=pyhd3eb1b0_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=py310h06a4308_0
tornado=6.2=py310h5eee18b_0
tqdm=4.65.0=py310h2f386ee_0
traitlets=5.7.1=py310h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.15=py310h06a4308_0
validate-email-address=1=pypi_0
vine=1.3.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.2.3=py310h06a4308_0
wheel=0.38.4=py310h06a4308_0
xz=5.2.10=h5eee18b_1
zeromq=4.3.4=h2531618_0
zlib=1.2.13=h5eee18b_0
zstandard=0.19.0=py310h5eee18b_0
zstd=1.5.5=hc292b87_0
